backend/requirements.txt
fastapi[all]==0.100.0
pydantic==2.5.0
pydantic-settings==2.1.0
xgboost==2.0.2
joblib==1.3.2
sqlmodel==0.0.14
python-multipart==0.0.6
uvicorn[standard]==0.24.0
pytest==7.4.3
ruff==0.1.6
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
requests==2.31.0

# New dependencies for chat system redesign
smolagents==1.20.0
litellm==1.74.3
websockets==12.0
redis==5.0.1
python-jose[cryptography]==3.3.0

# YAML template system dependencies
pyyaml==6.0.1
jinja2>=3.1.4

# Provider-specific dependencies (optional)
openai>=1.68.2
anthropic>=0.40.0 